requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain-google-genai
supabase
